feat(block-controller): refactor directory structure
