+++ src/scripts/block-controllers/index.ts
@@ -3,9 +3,16 @@
  * Author: Connor Doman
  */
 
-import ExpandingMediaController from "./expanding-media";
-import FeaturedVideoController from "./featured-video";
+import BlockController from "./block-controller";
+import ExpandingMediaController from "./misc/expanding-media";
+import FeaturedVideoController from "./misc/featured-video";
 import CyclingStackController from "./misc/cycling-stack";
 import NextProjectController from "./projects/next-project";
 
-export { ExpandingMediaController, FeaturedVideoController, CyclingStackController, NextProjectController };
+export {
+    BlockController,
+    ExpandingMediaController,
+    FeaturedVideoController,
+    CyclingStackController,
+    NextProjectController,
+};
