+++ src/scripts/block-controllers/misc/expanding-media.ts
@@ -3,9 +3,9 @@
  * Author: Connor Doman
  */
 
-import { getSiblings } from "../dom";
-import { BlockController } from "../dom/block-controller";
-import { randomIntInRange } from "../math";
+import BlockController from "../block-controller";
+import { getSiblings } from "../../dom";
+import { randomIntInRange } from "../../math";
 
 export default class ExpandingMediaController extends BlockController {
     static scrollThreshold: number = 150;
