+++ src/scripts/block-controllers/projects/next-project.ts
@@ -3,7 +3,7 @@
  * Author: Connor Doman
  */
 
-import { BlockController } from "../../dom/block-controller";
+import BlockController from "../block-controller";
 import { getImageColor } from "../../files";
 
 export default class NextProjectController extends BlockController {
@@ -11,7 +11,7 @@ export default class NextProjectController extends BlockController {
     private nextProjectImageBacking: HTMLElement;
 
     setup(): boolean {
-        this.debug = true;
+        this.debug = false;
 
         this.nextProjectImage = this.block.querySelector(".next-project-image > img") as HTMLImageElement;
         this.nextProjectImageBacking = this.block.querySelector(".image-backing") as HTMLElement;
